mtd: spi-nor: core: Introduce late_init()
Used to tweak various parameters that could not be extracted by
other means, i.e. when parameters are not defined in the JESD216
SFDP standard, or when the flash_info flags are incomplete.

As of now flash params are init statically via flash info flags, and
those parameters may be updated when parsing SFDP. Instead of overwriting
parameters at SFDP parsing time, it is desirable to init the flash only
by parsing SFDP, and if SFDP is not defined, to statically init the flash
via flash_info flags. We aim for the following pseudo code:

if SFDP is supported
	init flash params via SFDP
else if SKIP SFDP
	init flash params via flash info flags
else
	return error
call late_init()

Signed-off-by: Tudor Ambarus <tudor.ambarus@microchip.com>