ARM: dts: samsung: s5pv210: specify the SPI fifosize
Up to now the SPI alias was used as an index into an array defined in
the SPI driver to determine the SPI FIFO size. Drop the dependency on
the SPI alias and specify the SPI FIFO size directly into the SPI node.

The SPI nodes defined in s5pv210.dtsi are not enabled in any board file.
No SPI aliases are defined, thus I chose the FIFO size using common
sense: index 0 of the array corresponds to spi0, and index 1 to spi1.

Signed-off-by: Tudor Ambarus <tudor.ambarus@linaro.org>