firmware: exynos-acpm: move common structures to exynos-acpm.h
Prepare for the ACPM logging feature addition. ACPM is capable of logging
things to SRAM. The logging feature needs access to struct acpm_info
in order to get the sram_base, to the configuration data from SRAM, and to
the struct acpm_queue internal driver representation of a queue.
Move these structs to a common exynos-acpm.h.

Signed-off-by: Tudor Ambarus <tudor.ambarus@linaro.org>