nvmem: add Samsung Exynos OTP support
Add initial support for the Samsung Exynos OTP controller. Read the
product and chip IDs from the OTP controller register space and
register the SoC info to the SoC interface.

The driver can be extended to empower the controller become nvmem
provider. This is not in the scope of this patch because it seems the
OTP memory space is not yet used by any consumer, even downstream.

Signed-off-by: Tudor Ambarus <tudor.ambarus@linaro.org>